requirements.txt
aiomysql==0.1.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
idna==3.4
loguru==0.6.0
mysql-connector==2.2.9
numpy==1.24.2
pandas==1.5.3
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
win32-setctime==1.1.0
